ai-services/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
huggingface-hub>=0.15.0
elevenlabs>=0.2.26
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
numpy>=1.24.0
soundfile>=0.12.0
librosa>=0.10.0
aiofiles>=23.0.0
httpx>=0.24.0